Fix bugs found while end-to-end testing the HR module
Found by running the real backend/frontend against a local Postgres and
exercising every flow (department CRUD, employee CRUD, timeline events,
salary history, capo-limited permissions, org chart, analytics):

- Timeline events for cambio_reparto/cambio_capo stored raw numeric ids
  (from_value/to_value) instead of resolved names, so the UI would have
  shown "da 1 a 2" instead of department/manager names.
- No way to record a cessation date: data_cessazione was missing from the
  employee create/update schema; setting stato=cessato now defaults it to
  today if not given, and reactivating a cessato employee clears it again
  (fixed at the backend level, not just in the edit modal, so it holds for
  any client of the API).
- Empty optional text fields were saved as '' instead of NULL for every
  field except email, so a cleared field showed blank instead of "—".
- Postgres NUMERIC columns (salary, avg age/seniority in analytics) were
  serialized as strings by postgres.js, breaking currency formatting and
  risking bad chart scaling in Recharts; now cast to float8 in SQL.
- Duplicate department name returned a raw 500; now a clean 409 message,
  and the frontend's quick-add now surfaces backend errors instead of
  failing silently.
- Added a guard against a person being set as their own capo (previously
  would have silently vanished from the org chart tree).

+++ backend/src/modules/hr/analytics.ts
@@ -10,8 +10,8 @@ hrAnalyticsRoutes.get('/summary', requireModule('hr'), async (c) => {
   const [totals] = await db`
     SELECT
       COUNT(*) FILTER (WHERE stato != 'cessato')::int AS total_attivi,
-      ROUND(AVG(EXTRACT(YEAR FROM AGE(now(), data_nascita))) FILTER (WHERE stato != 'cessato' AND data_nascita IS NOT NULL), 1) AS eta_media,
-      ROUND(AVG(EXTRACT(EPOCH FROM AGE(now(), data_assunzione)) / (365.25*86400)) FILTER (WHERE stato != 'cessato'), 1) AS anzianita_media,
+      ROUND(AVG(EXTRACT(YEAR FROM AGE(now(), data_nascita))) FILTER (WHERE stato != 'cessato' AND data_nascita IS NOT NULL), 1)::float8 AS eta_media,
+      ROUND(AVG(EXTRACT(EPOCH FROM AGE(now(), data_assunzione)) / (365.25*86400)) FILTER (WHERE stato != 'cessato'), 1)::float8 AS anzianita_media,
       COUNT(*) FILTER (WHERE data_assunzione >= now() - interval '1 year')::int AS assunzioni_ultimo_anno,
       COUNT(*) FILTER (WHERE data_cessazione >= now() - interval '1 year')::int AS cessazioni_ultimo_anno
     FROM hr_employee
@@ -102,11 +102,11 @@ hrAnalyticsRoutes.get('/evolution', requireModule('hr'), async (c) => {
         WHERE e.data_assunzione <= (m.month_start + interval '1 month' - interval '1 day')
           AND (e.data_cessazione IS NULL OR e.data_cessazione > (m.month_start + interval '1 month' - interval '1 day'))
           AND e.data_nascita IS NOT NULL
-      ), 1) AS avg_age,
+      ), 1)::float8 AS avg_age,
       ROUND(AVG(EXTRACT(EPOCH FROM AGE(m.month_start, e.data_assunzione)) / (365.25*86400)) FILTER (
         WHERE e.data_assunzione <= (m.month_start + interval '1 month' - interval '1 day')
           AND (e.data_cessazione IS NULL OR e.data_cessazione > (m.month_start + interval '1 month' - interval '1 day'))
-      ), 1) AS avg_seniority_years
+      ), 1)::float8 AS avg_seniority_years
     FROM months m
     CROSS JOIN hr_employee e
     GROUP BY m.month_start
